Never use RDF+XML and reduce the range of options available for content negotiation.
This prevents running into this issue with RDF+XML:

https://www.w3.org/TR/REC-rdf-syntax/#section-Serialising

In particular some of our URIs on scotland look like this:

http://blah.com/blah/blah/blah/(something-in-parentheses)

This *is* a legal URI/URL and it will work with RDF+XML for anything
that isn't a predicate.  RDF+XML can't serialise predicates that contain
parentheses in the trailing URL path.  Madness!!!