Add handler for validating responses against a swagger specification.
Add a ring handler which validates the responses returned from an inner
handler against a swagger specification. The handler matches the route
of the incoming request to find the expected responses and checks the
body matches the expected schema for the response status if specified.

Remove swagger validations in draftset API routes.