Adjust delete-draft-graph-query so it works on stardog 4.1.2 and 4.2.3.
Problem is we're trapped between stardog versions.  4.1.2 is the most
stable/fastest/reliable, but prior to this commit tests that pass on
4.1.3 and above fail on 4.1.2.

The failure would result in too much state graph data being deleted
when a live graph l1 had two drafts d1 and d2 open on it.  when draft
d1 is deleted, the live graph l1 was also erroneously deleted.

The test suite has been rerun on stardog 4.1.2 and 4.2.3 and passes
under both versions.