Run the write loop on a dedicated thread.
Run the write loop on a dedicated named thread rather than running it
in a future. Use a flag to indicate whether the write loop should continue
rather than stopping the loop using future-cancel. Ensure the write loop
thread has stopped in stop-writer! - this will wait for the current job (if
any) to complete before returning.

Wait a short time for the write lock to be obtained when adding to the write
queue within queue-job!. The queueing thread holds the write lock for a short
period while adding a non-exclusive job to the queue which could cause other
job queueing attempts to be rejected.

Don't start the server when the repl starts - this causes a write-loop thread to
be started which can interfere with the write-loop created by some of the tests.