Ensure large SPARQL queries are sent as POST requests.
Send queries which would exceed Stardog's maximum URL length to be
sent in the body of a POST request. Stardog's maximum length is shorter
than the length specified in Sesame so override
SparqlSession.shouldUsePost to use the shorter threshold.

Add tests that GET/POST requests are sent based on the query length.