Prioritise exclusive-write over batch-write.  Fixes #161.
This fix is a bit of a hack to resolve an issue in PMD that may occur
since we've started copying graphs with sparql COPY statements rather
than manual batches.  Without this hack, setting the modified
timestamp data on a dataset can occur after the publication and
consequently not get applied to live.

Full write up on github issue #161.