Consume sequences generated from query operations eagerly.
When returning a lazy sequence generated from a streaming query result
(i.e. a tuple or graph query), the underlying HTTP connection is not
closed until the entire sequence is consumed or an error occurs when
fetching the next item in the sequence.

Add a query-eager-seq function which eagerly consumes the sequence
returned from grafter.rdf/query.

Ensure that query results are consumed immediately so HTTP connections
are not kept open longer than necessary. If multiple concurrent
requests are made to Drafter, it is possible that lazy sequences of
query results can cause the http-commons connection pool to be
exhausted (this is compounded by the small default pool size of 5 per
HTTP route). If a single operation holds multiple connections open
(as draftset-management/get-all-draftsets-by did before this change)
in lazy result sequences, it is possible for connections to be leaked
in the following scenario (assuming a connection pool size of 5 for
the stardog route).

1. Three concurrent requests are made and each successfully makes a
query returning a lazy sequence.
2. Each request tries to make a second request returning a lazy
sequence.
3. One of these requests will fail since the connection pool is now
full (3 initial requests + 2 of the second set of requests are held
open in the pool).
4. The failing request will throw a ConnecionPoolTimeoutException
and abort the request. The connection it initially created will not
be closed. The connection is being kept alive by the connection pool
and GCing the lazy sequence will not cause it to be closed either.

Ensuring all query result sequence are eagerly consumed should
prevent the above scenario.