Use .isLocked instead of taking and releasing the write lock.
Use ReentrantLock.isLocked to check whether an exclusive write is
in progress rather than taking the lock and releasing it. This means
that adding to the write queue no longer takes place under the lock
for non-exclusvie writes, although since exclusive writes do not
check the lock when queueing this is unlikely to have a noticable
impact.