Fix #160 better prioritisation policy of writes.
This policy should ensure only one actual write can occur on the
database at a time.  As before there are three types of writes (with
revised names), they are :blocking-writes (formerly :sync-writes),
:background-writes (formerly :batched-writes) and :publish-writes
(formerly :exclusive-write).

There are two concurrency primitives used to ensure that writes are
serialised.  A priority queue, served by a single thread; and a
global-writes-lock, providing mutual exclusion and coordination over
concurrent writes that we wish to prioritise (or quickly fail) as a
user will typically be waiting on them during a request.

Both :background-writes and :publish-writes are queued via the async
job queue.  :blocking-writes now skip this queue, allowing them to be
accepted whilst we are handling other writes.  Each of these classes
of writes can take global-writes-lock at certain times, depending on
the type of write they are doing, allowing other actors to either wait
for the lock, acquire the lock, or try for the lock with a timeout.

The following is a break down of the policy:

- Blocking Writes
  + Blocking writes happen concurrently with other sync writes (not queued
    on the async jobs queue).
  + Blocking writes wait for global-writes-lock for upto 10 seconds, if
    they wait longer than this before starting they will fail with a
    "writes temporarily unavailable message" if the lock is engaged.
  + Blocking writes will take the lock in order to execute one at a time
    with each other and other writers.  The main reason to do this is
    to allow them to wait up to 10 seconds if a batched-write has the
    lock.

- Background Writes
  + Are queued on to the async jobs queue.
  + appending / deleting of quads is batched and batches are
    interleaved with other batch writes jobs.  Whilst
    :background-writes are appending :blocking-writes will be
    scheduled between batches.
  + When a :background-write performs a graph copy it takes the
    global-writes-lock, for the duration of the graph copy only.
    Claiming the lock for these allows us to fail :blocking-writes
    fast, (if they are waiting for the lock to release for more than
    10 seconds, they will fail).

- Publish operations
  - :publish-writes are queed on to the async jobs queue.
  - :background-writes always happen before :publish-writes.
    (which is a hack to ensure PMD tasks occur in the right order,
    e.g. metadata updates happen before publish).
  - Publish jobs when started will no longer be split into joblets,
    and will occur entirely within the global-writes-lock.