Adjust implementation of JWT max-query-timeout override.
Prior to this change you could override the maximum to more than the
configured drafter default, but if you overode it to less than the
default you'd be allowed to request the default value, i.e. it would
be unclamped.

Change this for the simpler behaviour, to allow overriding the max and
getting what you asked for regardless of whether it was larger or
smaller than what was configured.

The `timeout` parameter is still used for the actual timeout you want,
this just lets you adjust the allowance if you're a trusted app.