Move PMD query timeout parameter to the request.
Look for the PMD-specific query timeout override in a
max-query-timeout parameter on the request. If present, it should
contain a MAC generated from the timeout value and the JWS signing
key shared between Drafter and PMD.

Refactor the sparql-timeout-handler middleware to take a function
(Request -> Timeout) instead of the endpoint timeout. The function
provided to the draftsets API and SPARQL endpoint handlers calculates
the timeout from the two request timeout parameters (timeout and
max-query-timeout) along with the configured endpoint timeout.

Split request timeout calculation into a chain of functions which
provide a timeout for different parts of the request. The chain is
[privileged timeout] -> [unprivileged timeout] -> [endpoint timeout] ->
[default timeout]. The privileged timeout handler is optional and is
only used if the DRAFTER_JWS_SIGNING_KEY environment variable is
configured.