Add test-generators namespace for generating initial database states.
Add test-generators namespace which allows tests to fully or partially
specify the initial state of the repository during tests. The statements
loaded into the test repository is randomly generated from the
specification.

Update tests in draft-management-test to use this namespace to generate
test data instead of setting it up imperatively.

Remove drafter.rdf.draft-management/append-data-batch! and replace it
with calls to grafter.rdf/add.