Support SPARQL queries in POST request bodies.
Issue #181 - Support the application/sparql content type for SPARQL
queries. The query parameter is passed unencoded in the request body in
such requests while GET requests pass parameters in the query string and
application/x-www-form-urlencoded POST requests pass them in the
URL encoded form.

If no content type is provided in POST requests, revert to the existing
behaviour of searching for the query parameter in either the query string
or URL-encoded form.

Update the swagger specification to require that SPARQL queries submitted
in formData should have content type of application/x-www-form-urlencoded.