Changes to how we run drafter tests.
Also added stardog ICV constraints in development to the drafter state
graph.  You can now run the tests with guard mode enabled in stardog
to prove that the state graph is "correct".

These constraints are now more necessary than before as new tests have
taken to loading the state from a turtle file, so we want to ensure
that these states are actually aligned with what the code elsewhere generates.

Added docs to README.  We now hardcode an assumption that
you have a database called drafter-test-db that is used for the
drafter unit tests.  In the future we should consider generating this
test database dynamically.