Record last access time on cache entries
This change resolves #242 and adds functionality to record and
retrieve last access times on cache entries.

The access times are recorded on entry into the cache (included by
default), and then on hit, and on lookup. The new function
`last-accessed-time` accepts a file path (not a cache key) and uses
this to lookup the time.