Report the job queue size as a gauge
This change replaces the drafter.jobs_queued metric with a
drafter.jobs_queue_size metric.

The former was being reported as a rate metric in DataDog, where rates
record as incidences per time-span, meaning that the decrements were
not being recorded correctly.

Due to our method of collecting metrics (library -> statsd -> datadog)
we don't have any way to record a metric that will be counted in
datadog as a count. Hence, the new metric is a gauge and will record
the current queue size at the points of insertion (including the
inserted job) and removal (not including the removed job).