Inject logging and datadog into all other components
The Logging and DataDog components both modify global state, and both
are required implicitly instead of explicitly by other components.

This change marks those implicit dependencies explicit by directly
configuring each component with a logger and a datadog. The components
don't have to do anything with these values that are passed in (and,
in fact, the values won't be anything useful), but Integrant can then
take advantage of these explicit dependencies to ensure that out
component initialisation order will always start with logging and then
datadog, and then everything else in its natural order afterwards.

This resolves #249 and also fixes the bug in #248

This also upgrades integrant due to a bug in 0.6.1 during the halt
process that would try to halt keys that are dependees of other
dependents, even if those dependees were not present in the system.