Replace core.cache stasher implementation with streams
This change removes the core.cache cache half implementation and
replaces it with a new cache that works by wrapping and intercepting
the handlers and result objects to insert things into the cache.

The file io has been hidden behind a backend that deals with io
streams only, hiding the file and disk implementation.

The tests have been maintained as much as possible, but they're a bit
muddled now, with the stasher-test ns still manually doing disk io for
example.

This commit also updates a bunch of edn files to reflect the changes
to the components