Bugfix: stop dropping cached boolean false results
When we call get-result it returns a nil to indicate that nothing was
found, which we then use to branch into running the query against the
real repo and storing the result.

Using this same mechanism for boolean results was causing the false
values to be understood as nothing-found, causing us to continue to
run the query and cache the result again.

This fix replaces our `or` usage with a `some?` on the result. This
returns true on all not-nil values, including falses, ensuring that we
branch correctly and then return the cached result.

This fixes bug #250