Bugfix: Include opts are in DataDog measure! calls
There is no name+body invocation of this fn, so our previous attempt
to do that was using our body as an opts arg, which caused it to be
invoked still, but without returning the result, and without timing
it.

i.e. not what we really intended.

This commit adds an opts are to each measure! call, meaning that we
now do return the result of the rest of the body, and time it
correctly while we do so.

This fixes bug #251