Don't cache state graph queries
This commit prevents caching state graph queries while still allowing
caching of other queries. It does so by inspecting the Dataset object
on the query objects and not caching
* If the graphs contain the state graph id
* Or if there is no dataset

The latter may catch non state graph queries and exclude them from the
cache, but we can't include them anyway without a time component in
the key, so that is fine for now.

A few tests have been commented out since their queries don't set any
graph restrictions, and we can't store anything in the cache without
it.

This resolves #253.