Replace rdf4j-repo with stasher-repo in the system
This replaces references to an instance of the rdf4j-repo in the app
with an instance of the stasher-repo instead. The `uncached-repo` and
`stasher-repo` keys both still exist, but they both point at the
stasher repo now.

This is a simple way to test that the stasher repo can handle all of
the requirements of the rdf4j-repo, and that the rest of the system
does not fall apart, before we go ahead and begin the disection of the
system and removal of the uncached repo.

This is part of #255 to remove the uncached repo completely.