Remove uncached-repo from the system*
*except the stasher internals that still use it.

This commit removes the uncached-repo and its usages from the entire
system. The `stasher-repo` key has also been removed and replaced with
a single `repo` key in the places that used it.

All call sites of uncached-repo or stasher-repo fns have been replaced
with the single repo.

With this commit, the entire system outside of stasher is cache
unaware and can occupy itself with just dealing with a generic "repo"
without caring about how it does what it does. Neat.

The internals of stasher require the uncached repo for the modified
times query. For now.

This is the bulk of development required for #255