Bugfix: Include the query cost in boolean miss metric
Boolean queries are eager, so by passing in an already realised result
we were only timing the cost of writing the result to disk, and not
the time of making the actual query. Oops.

This fix moves the metric to surround making the query and writing it
to disk.

This fixes #260