Add metrics to time uncached queries
Add timings to the queries that are not cached in Stasher, these are
those without a dataset or where the dataset is set to the state
graph.

This is intended to show us, as well as how long the queries take, the
amount of uncached queries that we make, and what the ratio is between
cached and uncached queries.

This resolves issue #257