Cache graph modified times
This commit adds caching of the graph modified times query based on a
state graph modified time of when-the-application-started (a new
juDate created on app startup).

There have been some problems caused by differences when reading tuple
results from StarDog vs reading them from the disk that was causing a
NPE in grafter. The bug was initially reported in #261 and has been
fleshed out in #264. This has been worked around in a branch on
grafter, for now, pending a real fix when we find out why the thing
that is happening is actually happening.

This partially resolves #258, but does not yet include cache
invalidation.