Enable caching for all state graph queries
This commit remove the default-turned-off of state graph query caching
and turns it on.

A `cache?` key has been added to the repo configuration that will
allow the entire cache to be turned off (state graph and non-state
graph queries), if we need to. I anticipate this being removed along
with the corresponding conditional branches after we are happy with
the cache timings.

This completes #259