Remove test that cannot pass
Now that the cache key can be made with a state graph modified time,
we have a problem in the tests that generate their own cache keys and
use them to find the files on disks.

The state graph modified time is stored inaccessibly in the
repo/connection, until we have it on the database it is not available
to be used here. And so, we can't use it to generate a key, and so we
can't find the files on the disk for the check.