Fix: NPE reading optional results from cache
We were seeing null pointer exceptions reading results from the disk
that used OPTIONAL in the query.

Grafter has been updated in grafter/#130 to handle nulls safely, which
prevents the exception, but does put nils in our result sets. Meaning
that the same query can have different results from the db (no nils)
or the disk (w/ nils).

I've added a test on the graph loading time to ensure everything
returns as it should. That's an OPTIONAL using query so a great
example.

This resolves #264