Bug, fix not reading stasher enabled env var
Setting the `STASHER_CACHE_ENABLED` env var was being ignored by the
app, with the cache always being turned on.

This was because we were placing in the opts as `enabled?` but then
looking for it under `cache?`.

This commit changes that to `cache?` in both places, and resolves the
issue #269.