Make Stasher thread pool configurable
This commit adds four new env vars that can be used to configure the
thread pool used for reading async cache results

The env vars and their defaults
* STASHER_CORE_POOL_SIZE      / 1
* STASHER_MAX_POOL_SIZE       / 10
* STASHER_QUEUE_SIZE          / 1
* STASHER_KEEP_ALIVE_TIME_MS  / 1000

With these defaults, the intial pool will contain a single
thread. When jobs come in and all threads are in use, a new thread
will be created up to a total of the max pool size (10).

When the pool size is greater than the core pool size (1), the extra
threads will be stopped if they are not used for > the keep alive
time (1000ms).

This commit also adds a beforeExecution hook that will report the pool
size to datadog for monitoring.

This commit fixes bug #275