Rename dev settings and start script files.
Rename config.env to dev-settings.env and start.sh to dev-start.sh
and specify their environment to be :dev and :ci in the package
manifest. Update the start script to fork the drafter process and
write a PID file.

Add a dev-stop.sh script for stopping the installed drafter if one
is running.