Improve error reporting when env vars are not set for SPARQL endpoints
IIRC there used to be better error reporting here, but as we've been
slowly moving from aero to integrant it appears to have got lost.  So
prior to this commit starting drafter without the appropriate env vars
being set would result in a rather opaque error.

This is a step towards a predominantly integrant based configuration
mechanism.