Make stasher tests more robust
Prior to this change intermittent test failures like the following
could happen:

FAIL in drafter.stasher.filecache-test/store-last-modified-on-cache-hit (filecache_test.clj:113)
Cache miss sets last access to now (more or less)
expected: (<= (->sec before-ms) (->sec last-access-ms) (->sec after-ms))
  actual: (not (<= 1563806759 1563806758 1563806759))
199 tests, 878 assertions, 1 failures.
The command "./travis/test-all" exited with 1.

This was because the cache was reused across tests, yet the tests
themselves assume this isn't the case.  So if the tests occur quick
enough it was possible for the above test to fail