Make timeout test more robust
It seems when we refactored the code base to use eager-query we subtly
changed the semantics of this test, as we were relying on the
connection blocking indefinitely.  Where as by switching to eagerly
consuming the results the connection would finish, and the test then
only worked by a coincidence of timings.