Delete redundant ->draftset constructor function
This shouldn't break any current consumers as defra-aims doesn't
appear to use this function.  Muttnik does but can easily be ported.

This implementation was removed as it was wrong, as it would coin
UUID's in various function arities; where as draftset ID's are only
strictly coinable by drafter.  So using this anywhere other than tests
or internally to this project would be a bug.