Define representation of the public endpoint
Issue #393 - Create an unmanaged graph containing a representation
of the public (i.e. live) endpoint. The endpoint is created on
startup if not already defined and contains the created and last
modified time of the live data. The updated time of the public
endpoint is set to the current time during draftset publish.

The JSON representation of the public endpoint contains id, type,
updated-at and created-at fields. A type field has been added to
the JSON representation of draftsets and these fields constitute
the Endpoint interface that both draftsets and the public endpoint
statisfy.

Add an /endpoint/public route which returns the JSON representation
of the public endpoint.

Add an /endpoints route for listing all endpoints visible to the
current user. This route supports the include parameter used by
the 'list draftests' route. The public endpoint is not owned or
claimable by any user so will not be returned if include is 'owned'
or 'claimable'.

Add a union-with-live parameter to the 'show draftset' route which
uses the latest modified time of the public and draftset endpoint
when set.

Add specs for the public endpoint and draftset records and remove
the existing prismatic schema definitions.