Prevent public endpoints graph being appended to
Issue #393 - Prevent the public endpoints graph from being created
inside a draftset through the append, delete or copy graph routes.
The public endpoints graph is not a managed graph so attempting to
delete from it is a no-op in the current implementation. Update
management/ensure-draft-exists-for to throw an exception if an
attempt is made to create a draft of the public endpoints graph.

Add check-endpoint-graph-consistent test macro which checks the
public endpoints graph is not modified during the test with the
possible exception of the updated time being modified.