Remove drafter.feature.endpoint.public/init component
Issue #393 - Remove the drafter.feature.endpoint.public/init
component which eagerly creates the public endpoint on startup if it
does not exist. Eagerly creating the endpoint causes the drafter
client tests to fail when checking for an empty database (the
default).

Update the check-endpoint-graph-consistent test macro to explicitly
create the public endpoint before each test.