Fix spec for Draftset record
Issue #393 - Move the email-string? spec from async.spec into util
and update it to use validate-email-address. Update the generator
to generate valid email addresses according to the stricter
validation function. Use this spec in the definition of the Draftset
spec and fix the draftset/create-draftset function to satisfy it.

Disable spec checking for endpoint route functions since these are
underspecified.