Handle missing public endpoint in list draftset and endpoint routes
Issue #393 - Update the list endpoints handler to remove the (nil)
entry for the public endpoint if it does not exist. Update the
show draftset handler to return the draftset record directly if
union-with-live is true but the public endpoint is missing.

Add test that publishing a draftset does not affect the public
endpoints graph if it does not already exist.

Avoid creating the public endpoint in check-endpoint-graph-consistent
and create it in each test that requires it. Change the check in
the macro to ensure no predicates are created in the endpoints graph.