Support union-with-live parameter on list draftset route
Issue #393 - Support the union-with-live parameter in the 'list
draftsets' route. If the flag is specified, merge each returned
draftset with the public endpoint in the same way as for the
show draftset route. This currently updates the last-modified time
to the latest modified time between the draftset and public endpoint.

Add union-with-live parameter to the get-draftsets operation in the
swagger specification and update the parameter's description.

Change the arguments to test get-draftsets function to represent
optional parameters with keyword arguments.