Add endpoint methods to drafter client
Issue #393 - Add endpoint namespace to drafter client containing
protocols for endpoints and endpoint references along with records
that implement them.

Add methods for the new drafter operations to fetch/create the
public endpoint and retrieve the list of all visible endpoints.

Add created-at and updated-at fields to the Draftset record and
add overloads of the ->draftset function for populating these
fields. Add test for from-json function.

Add optional options map parameter to the draftsets method supporting
the :include and :union-with-live parameters.

Add overload with optional parameters map to get-draftset method.
Add tests for specifying the union-with-live parameter.