Add handler for creating the public endpoint
Issue #393 - Support POST method in the public endpoint route which
creates the endpoint if it does not already exist. This route can
only be called by a user in the system role.

Add create-public-endpoint operation to the swagger specification
describing this operation.

Fix union-with-live parameter definition in get-draftset operation
specification.