Remove create-public-endpoint method from drafter client
Issue #393 - The create public endpoint route has been removed from
drafter so remove the corresponding method from drafter client. The
endpoint is now created automatically on startup if it does not
already exist and cannot be modified though the public API.

Modify the query to check for existing data in the test database
to ignore any statements in the public endpoints graph. Similarly
avoid deleting the endpoints graph when clearing test data at the
end of each test.