Replace capsule uberjar with skinny jar
Issue #432 - Change the JAR packaging for drafter from a Capsule
uberjar to a 'skinny' JAR which consists of an application JAR
along with a lib directory containing dependency JARs.

Add bin/pack script which builds the skinny JAR in the target
directory.

The skinny JAR is now added to the omni package as a zip containing
the appliation JAR and lib directory. Update the package manifest
to add this new item and extract it into in the install directory
during installation. Update the dev and production start scripts
to run drafter using clojure.main rather than via the uberjar.

Add bin/package script which builds an omni package for a skinny JAR
build in the target directory.

Update the travis build to run the pack and package scripts after
the tests pass. Publish the package files directly instead of
rebuilding in the deploy stage.

Remove the S3 and Github releases for drafter.jar since it can no
longer be executed standalone since it requires the contents of the
lib directory to run.

Remove the unused uberjar script and uber deps alias.