Only create the public endpoint automatically in development
Issue #393 - Attempting to create the public endpoint
automically in production creates a service dependency from drafter
to stardog. Defining this dependency within the systemd service is
not sufficient since stardog does not immediately accept requests
when started. Restarting stardog and drafter could therefore cause
drafter to fail to start if stardog is not available when the
endpoint initialiser component is run.

Remove the drafter.feature.endpoint.public/init component from the
main configuration and move it into init-public-endpoint.edn.
Specify this file in the integrant configutions within the
development start scripts so the endpoint is automatically created
in development.

Add :drafter.util/wait-for-connection component which polls for
and endpoint identified by a URI to become available within the
specified timeout. The public endpoint initialisation component
depends on and instance of this component to ensure stardog is
running before attempting to create the public endpoint.

The public endpoint must be created manually in production so add a
drafter.initialise namespace for running a collection of keys from a
custom integrant configuration specified on the command line.

Ensure the public endpoint is created when starting the test drafter
in the drafter client tests.