Add component for managing the creation of draft graphs
Issue #470 - Add drafter.backend.draftset.graphs namespace defining
a manager for the creation of managed and draft graphs. This is
configured with the set of protected graph specifications along
with the default http://publishmydata.com/graphs/drafter/.* graphs
which are always protected.

Remove the create-managed-graph!, delete-draftset-graph! and
create-draft-graph! functions from the draft-management namespace
and add corresponding function to the graph manager. The new
functions require the caller to specify whether the graph is
for protected or 'user' (i.e. non-protected) graphs. These check
that the graph is either protected or non-protected as required.

Update the append, delete, delete and copy graph jobs to use the
graph manager to create draft graphs. The set of protected graphs
is now only used by the graph manager so remove the
drafter/protected-graphs component from the integrant configuration.

Use the graph manager to check for protected graphs within the
draftset update query handler.

Remove conj-if macro and replace usages with cond-> macro and conj.