Implement missing methods for rewriting queries
Issue #464 - Implement the missing setBinding, removeBinding,
clearBinding, setDataset, getDatase and single-arguent evaluate
methods for rewriting graph and tuple queries. Query bindings and
graphs within datasets are subject to rewriting according to the
live->draft graph mapping. The reverse rewriting is done when
retrieving these values so the rewritten values are not visible
through the interface.

Rename the functions for creating rewriting queries within the
draftset.rewrite-result namespace to better indicate the return
types and add documentation.