Change implementation of draftset endpoints to use query interface
Issue #464 - Change the way queries are evaluated for draftset
endpoints to use the RDF4j query interface. The dataset for the
query is now calculated in terms of managed graphs and the
conversion to the underlying raw graphs is handled by the rewriting
query implementation.

Update the implementation of to-statements to restrict the queried
graphs according to the union-with-live? setting for the connection
and to rewrite results according to the graph mapping.

The process for calculating the dataset based on the user and query
datasets and the live graph restriction is now very similar for both
draft and live endpoints. Add a new rdf.dataset namespace containing
functions for converting Jena and RDF4j datasets to a clojure data
structure and for calculating the effective restriction based on the
incoming query and user datasets. Update the live endpoint to use
these to calculate its graph restriction.