Add Clock protocol for fetching the current time
Issue #486 - Add drafter.time namespace containing a Clock protocol
for resolving the current time. Create a system clock implementation
which returns the current UTC date/time and a test implementation
which can be updated manually. The protocol is also implemented for
the OffsetDateTime class which behave as fixed clocks which always
return the receiver.

Replace the time thunk arguments in various functions to take Clock
implmentations instead.

Add a drafter.time/clock component to the integrant configuration
and update the required data handlers to depend on the configured
clock.

Update the tests to use the test clock implementations where
necessary.

Add drafter.time.spec namespace containing specs for the clock
protocol and time representation.