Create interface for drafter operations
Issue #500 - Define an interface for each drafter operation and
create a test model Drafter implementation which uses the inner
implementation as much as possible. Separate the implementation
of each operation from the details of the HTTP API.

Add sync wrapper implementation which waits a configurable timeout
for each async operation.

Add drafter.manager component which contains the components required
by most operations - the backend repository, graph manager, global
writes lock and clock. Refactor each handler to depend on this
instead of each inner component separately.

Add constructor function for write locks and call it from the
integrant initialiser.

Refactor the delete-by-graph handler to separate the graph deletion
logic from the HTTP processing. The handler now calls a delete-graph
function which does the delete in a synchronous job and raises
an exception if the job fails or request is invalid.

Refactor update handler to allow UPDATE query strings to be parsed
directly.

Add specs for the write-scheduler, draftset.graphs and new
drafter.manager namespaces

Implement IDeref and IBlockingDeref on the Job record