Implement missing GraphQuery methods in 'all quads' query
Issue #477 - Implement missing methods for the GraphQuery returned
from operations/all-quads-query. Add backend.draftset.query-impl
namespace with a test-operation-methods function which tests each of
the methods defined by the RDF4j Operation interface behave as
expected. Use this to check the newly-implmented methods and add
tests that the 'all quads' and 'all triples' queries return the
expected results.

Update the tests within rewrite-result-test to use the assertions in
the query-impl namespace. Remove the previous functionality for
testing the Operation methods for rewriting queries.