Track graph state when translating UPDATE query operations
Issue #490 - Add a layer of abstract update operations to the
draftset update query handler which represent high-level operations
that are required to translate an update within the incoming SPARQL
UPDATE query. Update each UpdateOperation implementation to output
these operations instead of directly translating into Jena update
operations. Rename the raw-operations method of the UpdateOperation
protocol to abstract-operations to reflect this change.

Add an update plan structure which is updated in response to the
abstract actions output by each operation. The structure contains
the graph-meta map which is created at the start of the operation
containing information for each graph affected by the update.

Within a draft, a graph can either be unmanaged, exist only in
live or have an associated draft graph (managed and non-live is
possible if a graph is independently created within another draftset
but this case does not require special handling). The update plan
process now updates the state of a draft within a graph so later
operations do not unnecessarily create duplicate graph management
or draft entries in the state graph.

Add test assertions that check the state of modified graphs within
the state graph. These check there are no duplicate entries and the
managed/live state of the graph is as expected.

Fix the graph-meta query to explictly return the isPublic flag for
each live graph. Use the value of this flag to calculate the draft
state of each graph.

Add an update.spec namespace containing specs for the update endpoint
namespace.

Add a spec namespace for some of the functions and protocols in the
drafter.rdf.jena namespace.

Add type hints to remove all reflection warnings within the update
namespace. Temporarily comment-out the prepare-update function which
could be required in future.

Add generators for drafter quads and each component and use these
random values within each test to replace the ad-hoc quad
representation. Use the existing endpoint quad/triple queries to
check the contents of each draftset after each UPDATE query is
submitted.