Remove drafter/logging component
Issue #510 - Remove the unused drafter/logging component from the
base and test integrant configurations. Logging is now configured
with log4j2 directly and the logging-config.edn files are not being
used.

Update log4j2.xml to allow the location of the logs directory to be
specified by the DRAFTER_LOGS_DIR environment variable.

Remove all log-config.edn files.

Add a log4j2-test.xml logging configuration file based on the
test logging-config.edn definition. This logs to drafter-test.log
and sparql-test.log files within the log directory. This log
configuration will be used in preference to log4j2.xml defined in
the main resources directory if it is on the classpath such as when
running the tests or running the REPL with the :test alias. Like
the log4j2.xml configuration the location of the logs directory
can be overriden with DRAFTER_LOGS_DIR.